Move files around to avoid typescript-eslint error due to old deps https://github.com/typescript-eslint/typescript-eslint/issues/2446
